// src/saf/t/types.ts
export type SafTSourceDocumentType =
  | 'sales_invoice'
  | 'purchase_invoice'
  | 'payment'
  | 'movement_of_goods'
// ...
export type SafTPaymentMechanism =
  | 'CC' // Credit card
  | 'CD' // Debit card
  | 'CH' // Cheque
  | 'CO' // Cash on delivery
  | 'CS' // Settlement of credit balance
  | 'DE' // Electronic money
  | 'LC' // Letter of credit
  | 'MB' // SEPA / multibanco reference
  | 'NU' // Cash
  | 'OU' // Other
  | 'PR' // Permuta (barter)
  | 'TB' // Bank transfer
  | 'TR' // Restaurant / sodexo voucher